feat(day14): add function to drop sand and check if it has landed
This commit adds a new function to drop sand in the cave, given a starting
point and a set of rules to determine if the sand has landed or not. The
function returns the number of sand piles that have landed. It also updates
part1 and part2 to use this function instead of duplicate code.

Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>